It doesn't make sense to have floating point heart rate reading #353

+++ lib/devices/device_descriptors/data_handler.dart
@@ -112,10 +112,10 @@ abstract class DataHandler {
     return timeMetric?.getMeasurementValue(data);
   }
 
-  double? getHeartRate(List<int> data) {
+  int? getHeartRate(List<int> data) {
     if (heartRateByteIndex == null || heartRateByteIndex! >= data.length) return null;
 
-    return data[heartRateByteIndex!].toDouble();
+    return data[heartRateByteIndex!];
   }
 
   void clearMetrics() {
